+++ src/loader/endpoints.yaml
@@ -1,0 +1,11 @@
+- endpoint: br/cities/cases/full
+  python_file: get_cases
+  update_frequency_minutes: 10
+
+- endpoint: br/cities/embaixadores
+  python_file: get_embaixadores
+  update_frequency_minutes: 10
+
+- endpoint: br/cities/simulacovid/main
+  python_file: get_simulacovid_main
+  update_frequency_minutes: 10
