+++ src/loader/endpoints/get_{template}.py
@@ -1,0 +1,18 @@
+def now(config):
+    """This method is going to be called by main.py and it should return the output
+    DataFrame.
+
+    Parameters
+    ----------
+    config : dict
+    """
+
+    pass
+
+
+# Output dataframe tests to check data integrity. This is also going to be called
+# by main.py
+TESTS = {
+    "test1": None,
+    "test2": None,
+}
