+++ src/loader/scheduler.py
@@ -5,17 +5,25 @@ from main import main
 from logger import log
 
 
+def handler(endpoint):
+
+    try:
+        main(endpoint)
+    except Exception as e:
+        log(
+            {"origin": "Datasource Scheduler", "error_type": "Generic", "error": e},
+            status="fail",
+        )
+        print(e)
+
+
 def start():
 
     while True:
-        try:
-            main()
-        except Exception as e:
-            log(
-                {"origin": "Datasource Scheduler", "error_type": "Generic", "error": e},
-                status="fail",
-            )
-            print(e)
+
+        for endpoint in endpoints:
+
+            handler(endpoint)
 
         sleep(float(os.getenv("REFRESH_RATE_MINUTES")) * 60)
